requirements.txt
torch
torchvision
transformers
Pillow
scikit-learn
matplotlib
faiss-cpu
streamlit